pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labpilot"
version = "0.1.0"
description = "LabPilot hackathon MVP for experiment recommendation."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.2.0",
  "scikit-learn>=1.4.0",
  "numpy>=1.26.0",
  "joblib>=1.3.0",
  "openpyxl>=3.1.0",
  "matplotlib>=3.8.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "python-multipart>=0.0.9",
  "tavily-python>=0.7.10",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["modeling"]

